Add optional `extension` keyword argument to set_cover()
Previously, set_cover() would rely on the Python standard module
`imghdr` to determine the type of the specified cover image file.
Unfortunately, imghdr does not support SVG, so set_cover() could
not handle that type (which is permitted by the EPUB spec).

Detecting SVG files programmatically is nontrivial, and is generally
considered to require parsing the file source.

Instead, simply provide an option for the client to specify the file
type with the new keyword argument.