Add dcterms to package.opf template
The EPUB spec allows for Dublin Core metadata terms. This commit adds
support for them to the package.opf template.

No API change is necessary—all metadata is already passed through to the
template.

It is the client's responsibility to ensure that only valid DC
terms are supplied (just as page content HTML).

https://www.w3.org/publishing/epub3/epub-packages.html#sec-metadata-reserved-prefixes
https://www.dublincore.org/specifications/dublin-core/dcmi-terms/